+++ src/command/encode.rs
@@ -1,9 +1,10 @@
 use crate::{
-    command::{auto_encode, PROGRESS_CHARS},
+    command::{args, PROGRESS_CHARS},
     console_ext::style,
     ffprobe,
     process::FfmpegProgress,
-    svtav1, temporary,
+    svtav1::{self},
+    temporary,
 };
 use clap::Parser;
 use console::style;
@@ -15,20 +16,15 @@ use tokio_stream::StreamExt;
 /// Simple invocation of ffmpeg & SvtAv1EncApp to encode a video.
 #[derive(Parser)]
 pub struct Args {
-    /// Input video file.
-    #[clap(short, long)]
-    pub input: PathBuf,
+    #[clap(flatten)]
+    pub svt: args::SvtEncode,
 
-    /// Encoder constant rate factor. Lower means better quality.
+    /// Encoder constant rate factor (1-63). Lower means better quality.
     #[clap(long)]
     pub crf: u8,
 
-    /// Encoder preset. Higher presets means faster encodes, but with a quality tradeoff.
-    #[clap(long)]
-    pub preset: u8,
-
     #[clap(flatten)]
-    pub encode: auto_encode::EncodeArgs,
+    pub encode: args::EncodeToOutput,
 }
 
 pub async fn encode(args: Args) -> anyhow::Result<()> {
@@ -44,11 +40,10 @@ pub async fn encode(args: Args) -> anyhow::Result<()> {
 
 pub async fn run(
     Args {
-        input,
+        svt,
         crf,
-        preset,
         encode:
-            auto_encode::EncodeArgs {
+            args::EncodeToOutput {
                 output,
                 audio_codec,
                 audio_quality,
@@ -57,7 +52,7 @@ pub async fn run(
     bar: &ProgressBar,
 ) -> anyhow::Result<()> {
     let defaulting_output = output.is_none();
-    let output = output.unwrap_or_else(|| default_output_from(&input));
+    let output = output.unwrap_or_else(|| default_output_from(&svt.input));
     // output is temporary until encoding has completed successfully
     temporary::add(&output);
 
@@ -66,17 +61,15 @@ pub async fn run(
     }
     bar.set_message("encoding, ");
 
-    let probe = ffprobe::probe(&input);
-    let has_audio = probe.as_ref().map_or(true, |p| p.has_audio);
-    let duration = probe.as_ref().map(|p| p.duration);
-    if let Ok(d) = duration {
+    let probe = ffprobe::probe(&svt.input);
+    let svt_args = svt.to_svt_args(crf, probe.fps)?;
+    let has_audio = probe.has_audio;
+    if let Ok(d) = probe.duration {
         bar.set_length(d.as_secs());
     }
 
     let mut enc = svtav1::encode(
-        &input,
-        crf,
-        preset,
+        svt_args,
         &output,
         has_audio,
         audio_codec.as_deref(),
@@ -87,7 +80,7 @@ pub async fn run(
         if fps > 0.0 {
             bar.set_message(format!("{fps} fps, "));
         }
-        if duration.is_ok() {
+        if probe.duration.is_ok() {
             bar.set_position(time.as_secs());
         }
     }
@@ -98,7 +91,7 @@ pub async fn run(
 
     // print output info
     let output_size = fs::metadata(&output).await?.len();
-    let output_percent = 100.0 * output_size as f64 / fs::metadata(&input).await?.len() as f64;
+    let output_percent = 100.0 * output_size as f64 / fs::metadata(&svt.input).await?.len() as f64;
     let output_size = style(HumanBytes(output_size)).dim().bold();
     let output_percent = style!("{}%", output_percent.round()).dim().bold();
     eprintln!(